requirements.txt
aiosqlite
discodo
discord.py
discord-py-slash-command